requirements.txt
flask
flask-login
psycopg2
flask-bootstrap
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
flask-socketio